requirements.txt
bottle==0.12.8
redis==2.10.3
configparser==3.3.0.post2
requests==2.3.0
jsonpickle==0.9.2
boto==2.38.0
cherrypy==3.8.0
cassandra-driver==2.7.1
bcrypt==2.0.0
